Code Transform thiing
develop the Transform UI so you can only transform the character at a certain time

+++ Karel_Verhulst_NewMediaProject/Assets/Scripts/Character/ManaBehaviour.cs
@@ -28,7 +28,7 @@ public class ManaBehaviour : MonoBehaviour
         IsManaEmpty = false;
     }
     
-    public void TrySpendMana(int amount)
+    public void TrySpendMana(float amount)
     {
         //if (_manaAmount <= 1.5f)
         //{
